// src/app/models/equipments/parts.model.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Part {
    charging?: string;
    exclusive?: boolean;
    id?: number;
    idCompany?: number;
    idModel?: number;
    idSegment?: number;
    name?: string;
    picture?: string;
    provider?: string;
    serialNumber?: string;
    checked?: boolean;
}